+++ docs/usage/basket.rst
@@ -1,0 +1,4 @@
+.. basket_checkout:
+
+Basket
+======
