--- docs/usage/basket_checkout.rst
@@ -1,4 +1,0 @@
-.. basket_checkout:
-
-The Basket and Checkout API
-============================
