+++ docs/usage/payments.rst
@@ -1,4 +1,4 @@
 .. payments:
 
-Payment backends
-================
+Capturing Payments
+==================
