+++ longclaw/longclawcheckout/urls.py
@@ -3,13 +3,13 @@ from longclaw.longclawcheckout import api
 from longclaw.settings import API_URL_PREFIX
 
 urlpatterns = [
-    url(API_URL_PREFIX + r'payment/$',
+    url(API_URL_PREFIX + r'checkout/$',
         api.capture_payment,
-        name='payment'),
-    url(API_URL_PREFIX + r'order_prepaid/$',
+        name='checkout'),
+    url(API_URL_PREFIX + r'checkout/prepaid/$',
         api.create_order_with_token,
-        name='order_prepaid'),
-    url(API_URL_PREFIX + r'create_token/$',
+        name='checkout_prepaid'),
+    url(API_URL_PREFIX + r'checkout/token/$',
         api.create_token,
-        name='create_token')
+        name='checkout_token')
 ]
